pyproject.toml
[tool.poetry]
name = "packet-loss-rate"
version = "0.1.0"
description = ""
authors = ["esmail211 <smaeil0018@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyshark = "^0.4.5"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
